moved all non-routing methods from views to view_functions
